+++ noudar-rendering/VBORegister.h
@@ -5,7 +5,7 @@
 #ifndef DUNGEONSOFNOUDAR_NDK_VBOREGISTER_H
 #define DUNGEONSOFNOUDAR_NDK_VBOREGISTER_H
 
-using VBORegister = std::tuple<unsigned int, unsigned int, unsigned int>;
+using VBORegister = std::tuple<uint8_t, uint8_t, uint8_t>;
 
 using VBORegisterId = std::string;
